requirements.txt
easydict==1.9
Flask==1.1.4
Flask_Cors==3.0.10
lion_pytorch==0.1.2
matplotlib==3.3.3
networkx==2.5.1
numpy==1.19.2
PyYAML==6.0.1
scipy==1.5.3
torch==1.8.1
torch_geometric==1.7.2
torch_scatter==2.0.8
torch_sparse==0.6.12
tqdm==4.60.0
